add options for host directory and host URL
document paths are absolute (with respect to the domain), which avoids
dealing with relative paths betweendocuments.

now, if the site is hosted in a subdirectory, the document paths will be
adjusted accordingly instead of pointing to the host domain.

the host domain being configurable will probably turn out useful in the
future.