rearrange files
this is a huge and confusing change, but it doesn't really do much
except putting files more closely together semantically. the
refactoring isn't complete because the code doesn't lend itself to
that yet, primarily due to the templating pattern being needlessly
complicated. this needs to be addressed by migrating to the
pattern developed in <https://github.com/fricklerhandwerk/modular-templates>.

also it still needs to be figure out what to do with content types that don't
nessarily have a file representation, such as navigation trees or tables of
contents.